Add AbstractZLLSensorConfig
