+++ lib/SensorType/ZLLLightLevel/Config.js
@@ -1,14 +1,14 @@
 'use strict';
 
-let AbstractSensorConfig = require('../AbstractSensorConfig');
+let AbstractZLLSensorConfig = require('../AbstractZLLSensorConfig');
 
 /**
  * ZLLLightLevel sensor: Config
  */
-class Config extends AbstractSensorConfig {
+class Config extends AbstractZLLSensorConfig {
   /**
    * Get dark threshold
-   * 
+   *
    * @return {int} Dark threshold
    */
   get darkThreshold() {
@@ -17,7 +17,7 @@ class Config extends AbstractSensorConfig {
 
   /**
    * Set dark threshold
-   * 
+   *
    * @param {int} Threshold
    */
   set darkThreshold(threshold) {
@@ -26,7 +26,7 @@ class Config extends AbstractSensorConfig {
 
   /**
    * Get threshold offset
-   * 
+   *
    * @return {int} Threshold offset
    */
   get thresholdOffset() {
@@ -35,29 +35,11 @@ class Config extends AbstractSensorConfig {
 
   /**
    * Set threshold offset
-   * 
+   *
    * @param {int} Threshold
    */
   set thresholdOffest(threshold) {
     this.attributes.set('tholdoffset', parseInt(threshold, 10));
-  }
-
-  /**
-   * Get battery level
-   *
-   * @return {int} Battery level
-   */
-  get battery() {
-    return this.attributes.get('battery');
-  }
-
-  /**
-   * Get reachable
-   *
-   * @return {bool} True if reachable, false if not
-   */
-  get reachable() {
-    return this.attributes.get('reachable');
   }
 }
