+++ lib/SensorType/ZLLPresence/Config.js
@@ -1,11 +1,11 @@
 'use strict';
 
-let AbstractSensorConfig = require('../AbstractSensorConfig');
+let AbstractZLLSensorConfig = require('../AbstractZLLSensorConfig');
 
 /**
  * ZLLPresence sensor: Config
  */
-class Config extends AbstractSensorConfig {
+class Config extends AbstractZLLSensorConfig {
   /**
    * Get sensitivity
    *
@@ -22,60 +22,6 @@ class Config extends AbstractSensorConfig {
    */
   get maxSensitivity() {
     return this.attributes.get('sensitivitymax');
-  }
-
-  /**
-   * Is user test?
-   * 
-   * @return {bool} True if user test mode, false if not
-   */
-  get userTest() {
-    return this.attributes.get('usertest');
-  }
-
-  /**
-   * Set user test
-   * 
-   * @param {bool} value True or false
-   */
-  set userTest(value) {
-    this.attributes.set('usertest', Boolean(value));
-  }
-
-  /**
-   * Get pending config
-   * 
-   * @return {array} List of configs pending change on device
-   */
-  get pending() {
-    return this.attributes.get('pending');
-  }
-
-  /**
-   * LED indication?
-   * 
-   * @return {bool} True if LED indication, false if not
-   */
-  get ledIndication() {
-    return this.attributes.get('ledindication');
-  }
-
-  /**
-   * Set LED indication
-   * 
-   * @param {bool} value True or false
-   */
-  set ledIndication(value) {
-    this.attributes.set('ledindication', Boolean(value));
-  }
-
-  /**
-   * Get battery level
-   *
-   * @return {int} Battery level
-   */
-  get battery() {
-    return this.attributes.get('battery');
   }
 }
