+++ lib/SensorType/ZLLTemperature/Config.js
@@ -1,10 +1,10 @@
 'use strict';
 
-let AbstractSensorConfig = require('../AbstractSensorConfig');
+let AbstractZLLSensorConfig = require('../AbstractZLLSensorConfig');
 
 /**
  * ZLLTemperature sensor: Config
  */
-class Config extends AbstractSensorConfig {}
+class Config extends AbstractZLLSensorConfig {}
 
 module.exports = Config;
